requirements.txt
qrcode
pillow
python 3